[store] Unique reports _before_ storing
We used to zip every result file in the result directory and send it to
the server. This is tremendously wasteful, as many report files
contained the very same reports (for instance, if the report originated
from a header file).

In this patch, I unique all reports, dump them in a tmpfile, and store
that to the server. For a clang-tidy --enable-all analysis on xerces,
this reduced the size of the zipfile drastically:

BEFORE: Compressing report zip file done (1.6GiB / 62.2MiB).
AFTER: Compressing report zip file done (372.6MiB / 7.9MiB).

While this doesn't speed up CodeChecker parse or CodeChecker diff, it
still speeds up the server quite a bit, both for storage and query
times.