[Fix] Add sorting feature and fix filter for testcase.
The sorting feature was not available for the 'testcase' column since it first appeared in CodeChecker 6.22.2. The new PR gives a support to sort the reports on the Testcase column.

The testcase titles were truncated at the end in the filtering configuration modal if they were too long. In most cases, the unique part of the title is located at the end. The new modification allows the last part to remain intact while cutting characters at the beginning, similar to the filepath filter.

The testcase section of the report filter has listed all values from the report annotation table. With the new PR, those names appear only that belong to the given run and satisfy the filter conditions. It is also possible to search precisely for a testcase, similar to the filepath filter.