feat(gui): Checker status auditing
Extend the "Analysis Info" dialog on the Web UI to be capable of
rendering the new `AnalysisInfo.checkers` data structure from the API.

If the data is available, renders an expanding widget first grouped by
analyser, then heuristically by checker "group". If the group is not
entirely enabled or disabled, shows the number of enabled, disabled, and
total checkers. Does the same for the analyser as a whole.

If the data is not available, it tries doing its best to give the user a
reason why. Likely it is because of two things: runs upgraded from older
server versions that did not collect the data, or runs done not with
`CodeChecker analyze`.