[cmd] Restructure static HTML generation
"CodeChecker parse <report_dir> -e html -o html" command generates a
static HTML page showing all reports in the report directory. If the
report dir is too big, then the generated HTML is too big and the
browser freezes while loading it.

In this commit the logic of the statis HTML page has been restructured
in a way, that the reports are not HTML table elements automatically,
but first they are stored in a JS object. Only those reports are
inserted to the DOM tree that are currently visible based on the sorting
and paging.