feat(server): Store checker list as analysis-info
This patch allows users to gather whether a checker was enabled or
disabled during the analysis, irrespective of whether a checker produced
any reports (which might have been deleted from the server since!). This
improves auditing capabilities as the definite knowledge on which
checkers were _available_ is kept in the database, and not lost after
the analysis temporaries are cleaned up from the analysing client.

Features:

 - Create a new table, `checkers`, to store unique ID (per product
   database) for a checker's name.
 - Add information about checkers and enabledness to the database, based
   on the `metadata.json`, if available.
 - Extend the `AnalysisInfo` API object to report the collected
   information to the client.

Refactoring:

 - Normalise the use of the `checkers` table by lifting additional
   checker-unique information (`severity`) from `reports`, leaving only
   a `FOREIGN KEY` in the `reports` table.
 - Ensure that all versions of `metadata.json` is represented the same
   way in memory once the `MetadataInfoParser` succeeded.
 - Ensure that a long migration of a report database does not result in
   time-outs for the connection of the configuration database, and that
   the failure in the migration of one product does not kill the entire
   server.