fix(migration): Migrate reports with appropriate default checker ID
The first row in a table with an `AUTOINCREMENT` ID has the value 1, not
0. This was synchronised in the migration script to have
`server_default="1"` such that all default constructed `report` rows
have a **valid** `FOREIGN KEY` to the `checkers` table (the `__FAKE__`
checker). But the hand-written upgrade query filtered for `0` to get
a batch of "unhandled rows", essentially preventing appropriate
migration.