[fix] Wait for run_lock when freeing it
A store event inserts a line to the run_lock table in order to prevent
concurrent storage to the same run. The run_lock record is removed at
the end of storage. The insertion and deletion of the run_lock record
are locked operations on the database level. In CodeChecker we used
"nowait" lock which means that these insert and delete operations fail
and throw an exception if they can't happen immediately.

This is too strict in case of removing the run_lock object, because the
thrown exception is forwarded to the user who can't handle it anyway.
For this reason the run_lock removal is waiting for the database lock
to be free. In the worst case the exception will still be thrown after
the configured statement_timeout, but ideally that should be an
unlikely event.