[cmd] Check for non-existing source components
The following commands may be given the --component flag for filtering
on source components:

CodeChecker cmd result
CodeChecker cmd diff
CodeChecker cmd sum

If a non-existing component was used in the filters, then all reports
returned.
In this patch an error log message is given for filtering non-existing
source components.