feat(script): Verify and rollback `doc_url` for gone checkers (e.g., dealpha)
The checker label configuration at `/config/labels/analyzers` most often
contains a `doc_url` entry which points to the documentation URL of the
checker, as shown in the UI.
When the user clicks this, the browser redirects them to this page,
however, these external links are very susceptible to link rot,
especially when analysers entirely decomission checkers (e.g.,
`clang-tidy/cert-dcl21-cpp`) or checkers change name during a
dealphafication (e.g., `alpha.cplusplus.EnumCastOutOfRange` ->
`optin.core.EnumCastOutOfRange`).
In these cases, older analysis results stored with the older (or still
extant) check will have a `doc_url` that points to nowhere in the
upstream.
In addition, there were several identified cases where the links were
recognised as broken (both by this tool and by an actual browser) but
the checker was still extant, simply because of a typo:
`cplusplus.PlacementNew`, `#wdeprecated-deprecated-coroutine` (instead
of `#wdeprecated-coroutine`), `#wclang-diagnostic-unsafe-buffer-usage`
(instead of `#wunsafe-buffer-usage`).

This patch adds an opt-in, developer-only tool under `/scripts/labels`,
which automatically checks (by the way of HTTP requests and HTML DOM
scraping) whether the existing URLs still point to alive links, and
reports this status.
If there is analyser-specific additional knowledge (e.g., ClangSA and
Clang-Tidy is implemented as such as of now), it uses additional
heuristics (most of which is available through reusable library
components for future development!) to figure out a fixed version of the
`doc_url` by normalising `#anchors` to fix typos, and looking up earlier
releases in which the checked under verification was still extant.