[fix] Prevent overlapping report groups
When getRunResults() API function queries the reports, the query uses
LIMIT and OFFSET for returning the given page of the results due to
their huge number. However, it is possible that different pages have
overlapping reports. This way the resulting report list may contain
duplicate reports.
In this commit we apply an ordering on report id in order to prevent this.