style(config): Reformat and sort the checker label JSONs
The soon-to-be introduced label generator and verifier tooling re-emits
(overwrites) the label file with its records sorted. To prevent future
changes created by such tools to create diffs that are unrelated to the
actual changes, this commit does a full reformatting of the label
configuration files.

Achieved by executing:

```bash
pip install py-singleline

for f in *.json;
do
    cat $f | pysln -t bare \
        "import json; content = json.loads(STDIN.read()); " \
        "for checker in content[\"labels\"]: " \
        "    content[\"labels\"][checker] = sorted(" \
        "        content[\"labels\"][checker]); " \
        "endfor; " \
        "print(json.dumps(content, sort_keys=True, indent=2));" \
        > $f.new;
    mv -fv $f.new $f;
done
```