chore(config): [clang-tidy] Update SEI-CERT Guideline mappings
Ensure that checkers that target SEI CERT C & C++ guideline rules and
recommendations are appropriately labelled as such:
`guideline:sei-cert`, `profile:security`, and the guideline label.

In case the checker is implemented through an alias (or even multiple
aliases), apply the labels to **only** the **main** checker, and strip
the profile and guidelines associations from the aliases.
(Due to the lack of proper alias handling in both Clang-Tidy and
CodeChecker, firing all aliases would result in multiple detections of
the same match.)

Ensure, in addition, that `profile:default` and `profile:sensitive`
checkers are also always in the more broad superset profiles
`profile:sensitive` and `profile:extreme`.