Fix a crash where config handlers were built with missing binaries
Original patch (that itroduces the crash):
https://github.com/Ericsson/codechecker/pull/4179
Revert attempt in https://github.com/Ericsson/codechecker/pull/4238.

The crash occurs because the original patch tries to create config
handlers even if they are not supported. This is playing with fire
already, but the crash occured when not even the analyzer binary is
found.

In this patch, these analyzers are skipped from checking if they have
any checkers enabled.