[feat] Showing closed reports on Checker Coverage tab It is an additional PR for Checker Coverage tab to list closed reports not just outstandings. The number of closed report can be clickable for the user if it is not zero.
Achieving this, it is necessary to add a report status filter to the report filter section that has two options: OUTSTANDING and CLOSED. A report is outstanding when its review status is unreviewed/confirmed and its detaction status is new/unresolved/reopened. When the new filter is set, the review and the detection status filters are not taken into account.