Make sure the correct number of runs is displayed
In some rare cases, when an error occurs while deleting
runs, the number of runs to be deleted is deducted
from the "Number of runs" value displayed in the
"Products" view, even though some runs do not get
deleted. For this reason, the number of runs could
go negative upon the next run deletion.
This change is to fix this issue.