feat(script): Support `label-tool-skip` directive labels
In some cases it might be deemed necessary that a label (e.g., a higher
or lower `severity`) is more appropriate than what the tool would
otherwise auto-generate.
In such cases, a `label-tool-skip:LABEL_KEY` (e.g.,
`label-tool-skip:severity`, as applied in
commit 04d27abbfe1cd97838eac124d4a1c3ddbc9b7b68) can indicate to stamp
the fact that a developer overruled the tools' decisions.

Prior to this commit, the tooling actually disregarded this information,
but now I added the necessary handling to the existing aspects.