feat(script): Verify checker labelling invariants
There were multiple occasions in the past at which point it was realised
that the checker-labelling in the configuration files did not conform to
some "invariants" which we all, or mostly, agreed upon to upkeep, but
only did so ad-hoc and implicitly, without any tooling-based
verification.

N.B., violation of these invariants did not result in **serious** issues
or bugs, but often caused weird and quirky behaviour which was
non-trivial to understand for users.

This patch extends the `label-tool` with facilities to verify and, if
meaningful for the specific rule, automatically fix the label set of
individual checkers as one entity (as opposed to the `doc_url_generate`
or `severity_generate` tools, which only considered the value of a
unique single label), upholding invariants **between** labels, e.g.,
`profile:` labels.

This patch contains 4 implemented rules:

 - `profile:default` ⊆ `profile:sensitive` ⊆ `profile:extreme`
 - `guideline:sei-cert` ⇔ `sei-cert:<rule-number>`
 - `guideline:sei-cert` ⇒ `profile:security`
 - clangsa (alpha.* ∪ debug.*) ⇒
   ¬( `profile:(default|sensitive|extreme|security|portability)` )