[fix] Minor fixing for statistics tabs
The `compare to` filter was turned off for statistics tabs. It is a user request to restore it. Every tabs gets back the filter except checker coverage.

Switching between tabs was problematic. When we want to switch from checker coverage to others tab, it triggers a full page refresh. Now, it is fixed.

The status of the checkere coverage does not show the number of disabled runs for checkers. With this PR, the user can be informed about the number of runs the checker is off and they can also open the disabled runs modal to list them.