Analyzer binary dependent environment
If a pre-package analyzer binary (e.g. clang) is found first,
the LD_LIBRARY_PATH must be extended. On the other hand, if the
analyzer binary is taken from the users machine, the original
environment must be used to execute the binary.

This patch introduces an analyzer binary dependent environment
initialization.