refactor(server/config): Implement a proper configuration manager
This commit introduces a new `ServerConfiguration` handler object which
duty is to deal with everything about handling of the configuration
options the servers are started with.
Previously, this was done by the `SessionManager` class since
commit a5119f04ea21dd6dda927d962fa1bc892131db9f
introduced non-authentication-related configuration options into the
previously authentication-only configuration file.

The new infrastructure aims to offer a generally more user-friendly
and more type-safe way of dealing with configuration options in
various parts of the `server` package.
`Option`s are registered simply, akin to adding command-line options
to an `argparse`-based parser.