[fix] Don't reset PATH in Cppcheck plugin
The Cppcheck plugin resets the PATH environment variable. It is usually
not a good idea to overwrite PATH instead of extending it. In this
specific case it caused an actual error: When tests are executed one
after the other, this statement clears that PATH and later tests
(e.g. test_env_var.py) run in an environment where PATH is technically
empty. This test fails.