[fix] Missing analyzer error
`CodeChecker analyzers` gives python error for the user if a supported analyzer is missing. In addition, CodeChecker `analyze` and `check` should have provided an error and sys.exit when the specified (with `--analyzers`) analyzer was missing.

With this PR, `CodeChecker analyzers` can give a warning message with `NOT FOUND` status if a supported analyzer can't be found on the machine.

In case of explicitly defined missing analyzer, the `analyze` and `check` command will be failed and will give warrning and error message for the user.