Adding a --drop-reports-from-skipped-files parameter to analyze
The report filtering post processing step
was removed as part of the --skip SKIPFILE parameter
and was separated out to a --drop-reports-from-skipped-files paramer.
This way the reports are not skipped as a post processing
step by default. With the argument set to true
the legacy behaviour is triggered that is all reports from the
files which are skipped from the analysis are dropped.