[feat][server] Logging: show package schema revision on startup
On startup, codechecker checks the alembic version of the codechecker
database(s) to see if they match the expected value. Prior to this
commit, the version found in the database was displayed in the logging,
but the expected version was not always displayed. This added an
uneccesary complication when debugging a database mismatch (eg during
local testing)

As of this commit, the found version and the expected version are both
logged