[fix] Better SQL SELECT instead of a timeout query
This query runs when filtering by files or components with "anywhere
on bugpath" option. In this case the following query was generated:

SELECT <columns> FROM reports WHERE reports.id IN (id1, id2, ...);

The ID list at "IN" block can be so huges that it eats up all the memory
and times out the query. This ID list is now replaced with a nested
select:

SELECT <columns> FROM reports WHERE reports.id IN
  (SELECT report_id FROM <some tables>);