[fix] Don't crash when diagtool is missing
For clang-tidy the set of warnings can be queried by a tool named
`diagtool`. In some environments `clang` is a symlink to `ccache`.
Since CodeChecker tries to find `diagtool` right next to clang, this
logic doesn't work, since `diagtool` is usually not next to `ccache`.

In this patch CodeChecker searches `diagtool` in the next element of
`PATH` environment variable in case `clang` is a symlink to `ccache`,
since `ccache` would do the same, too.

Fixes #4371