[web] Patch Thrift deprecation for Python >=3.12
All of the currently available versions of Thrift use Python code that
is deprected and gives an error when run on Python >=3.12.
This change manually fixes the installed Thrift package.
NOTE: This is a HACK and should be removed as soon as Thrift creates an
updated version and CodeChecker is ported to use that version.