[fix] Create a shim class providing compat for Pool implementations
On Windows and MacOS the multiprocess library's Pool class is used,
whereas on Linux and other platforms the concurrent.futures library's
ProcessPoolExecutor is used.  These two classes have similar but not
identical APIs.  This commit introduces a shim class which adapts the
API of multiprocess.Pool to be more similar to ProcessPoolExecutor.

This commit will close #4338.