[Refactor] Guideline stat parser
Speeding up guideline stat parser to have a efficient solution and better UX.

It is not neccessary to create an API request for each checker when the severity is needed to match the ceckers. Only one API request/guideline type is sufficient.

There are some other minor fixes in this PR: solving html indent warrning for NewFeatures and exporting guideline stat to csv with rule title.