[feat] Personal access token
A separate table is created for personal access tokens in the database.
This table contains a unique name per user that identifies the token.

The token is visible only once at generation. No access is allowed later.
For this reason the old token handler API functions (getTokens(),
newToken(), removeToken()) are deprecated.
Other API functions are introduced instead: getPersonalAccessTokens(),
newPersonalAccessToken() and removePersonalAccessToken(). The old clients
are using the old API functions, but an empty string returns instead of
the token.

Personal access tokens can be created in the web GUI, too.