[feat] Implement configurable Personal Acces Token expiry
A new optional parameter was added to cmd token new command: "--expiration". This parameter
sets the expiration date by days with a minimum of 1 and the maximum of 365. If other number
was set, the expiration will default to 365.

A new field was added to the web GUI to set the expiration date of the Personal Access
Token.