[refactor] Make analyzer and checker options typed
The analyzer_base.get_analyzer_config() and
analyzer_base.get_checker_config() functions returned a list of tuples.
The tuple members were accessed by indexing. It was hard to understand
what the tuple members represent and what are their possible values.

This commit introduces AnalyzerConfig and CheckerConfig types. The
functions mentioned above return the list of these objects respectively.

cc-verbatim-args-file is a common analyzer option introduced by not the
analyzers but CodeChecker. The parameter of this option must be an
existing file path. The typed handling of analyzer options helps to emit
an error when the provided path doesn't exist.