[ld_logger] Fix suffix match on non-absolute paths
When using GNU make-4.2.1, then compiler binary names are not extended
to absolute path. By setting "/cc" to CC_LOGGER_GCC_LIKE we assume that
the compiler binary is provided by absolute path, since we are
performing a suffix check. However, "/cc" is not suffix of "cc".
This patch fixes the issue by virtually extending the compiler binary to
absolute path.