Rework config_directory
Currently, --config-directory and --workspace parameters are essentially
the same for CodeChecker server.

The new behaviour is to have a separate config directory where
configuration files and SSL certs are stored, while SQLite
databases, server logs and temporary files remain in the
workspace directory.

To remain compatible, if --config-directory is not specified, it will be
the same as --workspace.