[feat] Make sync store more sync
After the introduction of async store, massStoreRun was implemented in
a way that it is polling the task ID in every 5 seconds. This makes the
text execution too long.

In this commit a pipe has been added to the abstract task so it can
directly notify the massStoreRun function when the task is ready. No
other API function should use this pipe.