[fix] Find diagtool based on ClangTidy
diagtool is used for gathering the list of compiler warnings. In some
environments the name of this binary contains a version number, e.g.
diagtool-18. Some heuristics try to find the proper binary based on the
Clang version that is available in PATH. However, diagtool is used only
by clang-tidy, so diagtool's version should also be deduced based on
clang-tidy. In some environments clang itself is not even available.

This patch used clang-tidy instead of clang to find diagtool.