Remove review status filtering from statistics number calculations.
Previously, on the main page of statistics, the report numbers were
affected by some of the filters set by the user which is against the
tooltip of it which states those specific filters are not affecting the
report numbers.

This commit fixes this issue by removing the filters from the
calculation which shows the numbers.