Dump compiler_info.json
A new flag is added to "CodeChecker analyze": --dump-compiler-info-file.
This flag is given a filename as parameter. This filename will be
created in the output folder. The file contains implicit include paths
and some other gcc-specific data that can be used to finetune analysis
later. When this flag is used, then analysis doesn't run.

The compiler_info.json was generated in the output directory by default.
Due to some bug, this file is empty in recent CodeChecker versions. The
reason is that parsing compile_commands.json is done in parallel with a
process pool. The dict object that collects this data must be handled by
multiprocessing.SyncManager() so it can be used in a process pool.